# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aozora - Dorm Finder API for Lopez, Quezon
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/submit-verification:
    post:
      operationId: submitVerification
      tags: [users]
      summary: Submit ID verification request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVerificationRequest"
      responses:
        "201":
          description: Verification submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRecord"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get a public user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"

  # --- DORMS ---
  /dorms:
    get:
      operationId: getDorms
      tags: [dorms]
      summary: List dorms with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: minRent
          in: query
          schema:
            type: number
        - name: maxRent
          in: query
          schema:
            type: number
        - name: hasAvailableBeds
          in: query
          schema:
            type: boolean
        - name: amenities
          in: query
          schema:
            type: string
            description: Comma-separated list of amenities
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, rating_desc, newest]
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radiusKm
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of dorms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DormsListResponse"

    post:
      operationId: createDorm
      tags: [dorms]
      summary: Create a new dorm listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDormRequest"
      responses:
        "201":
          description: Dorm created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dorm"
...
  /dorms/{dormId}:
    get:
      operationId: getDormById
      tags: [dorms]
      summary: Get dorm details
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dorm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DormDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateDorm
      tags: [dorms]
      summary: Update dorm listing (owner only)
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDormRequest"
      responses:
        "200":
          description: Updated dorm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dorm"

    delete:
      operationId: deleteDorm
      tags: [dorms]
      summary: Delete dorm listing (owner only)
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /dorms/{dormId}/photos:
    post:
      operationId: addDormPhoto
      tags: [dorms]
      summary: Add a photo URL to a dorm listing
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPhotoRequest"
      responses:
        "201":
          description: Photo added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DormPhoto"
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get current user's favorite dorms
      responses:
        "200":
          description: Favorite dorms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DormsListResponse"

    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a dorm to favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteRequest"
      responses:
        "201":
          description: Added to favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /favorites/{dormId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a dorm from favorites
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed from favorites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /favorites/check/{dormId}:
    get:
      operationId: checkFavorite
      tags: [favorites]
      summary: Check if a dorm is favorited
      parameters:
        - name: dormId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Favorite status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteCheckResponse"

  # --- ADMIN ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # --- USER SCHEMAS ---
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, owner, admin]
        verificationStatus:
          type: string
          enum: [unverified, pending, verified, rejected]
        isSuspended:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, fullName, role, verificationStatus, isSuspended, createdAt]
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        role:
          type: string
        verificationStatus:
          type: string
        avatarUrl:
          type: string
          nullable: true
      required: [id, fullName, role, verificationStatus]
...
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
    SubmitVerificationRequest:
      type: object
      properties:
        idImageUrl:
          type: string
        idType:
          type: string
          description: e.g. "National ID", "Passport", "Driver's License"
      required: [idImageUrl, idType]
    VerificationRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        idImageUrl:
          type: string
        idType:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewNote:
          type: string
          nullable: true
        submittedAt:
          type: string
        reviewedAt:
          type: string
          nullable: true
      required: [id, userId, idImageUrl, idType, status, submittedAt]

    # --- DORM SCHEMAS ---
    Dorm:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: string
        monthlyRent:
          type: number
        address:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        amenities:
          type: array
          items:
            type: string
        totalRooms:
          type: integer
        bedsPerRoom:
          type: integer
        availableBeds:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        coverPhotoUrl:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        createdAt:
          type: string
      required: [id, ownerId, name, description, monthlyRent, address, amenities, totalRooms, bedsPerRoom, availableBeds, status, totalReviews, createdAt]
    DormDetail:
      allOf:
        - $ref: "#/components/schemas/Dorm"
        - type: object
          properties:
            owner:
              $ref: "#/components/schemas/PublicUser"
            photos:
              type: array
              items:
                $ref: "#/components/schemas/DormPhoto"
            isFavorited:
              type: boolean
          required: [owner, photos, isFavorited]
    DormsListResponse:
      type: object
      properties:
        dorms:
          type: array
          items:
            $ref: "#/components/schemas/Dorm"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [dorms, total, page, totalPages]
    CreateDormRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        monthlyRent:
          type: number
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        amenities:
          type: array
          items:
            type: string
        totalRooms:
          type: integer
        bedsPerRoom:
          type: integer
        availableBeds:
          type: integer
        coverPhotoUrl:
          type: string
      required: [name, description, monthlyRent, address, totalRooms, bedsPerRoom, availableBeds]
    UpdateDormRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        monthlyRent:
          type: number
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        amenities:
          type: array
          items:
            type: string
        totalRooms:
          type: integer
        bedsPerRoom:
          type: integer
        availableBeds:
          type: integer
        coverPhotoUrl:
          type: string
    DormPhoto:
      type: object
      properties:
        id:
          type: integer
        dormId:
          type: integer
        url:
          type: string
        caption:
          type: string
          nullable: true
        order:
          type: integer
      required: [id, dormId, url, order]
    AddPhotoRequest:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        order:
          type: integer
      required: [url]
...
    FavoriteRequest:
      type: object
      properties:
        dormId:
          type: integer
      required: [dormId]
    FavoriteCheckResponse:
      type: object
      properties:
        isFavorited:
          type: boolean
      required: [isFavorited]

    # --- ADMIN SCHEMAS ---
...