# lib/api-zod/src/generated/types/updateDormRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aozora - Dorm Finder API for Lopez, Quezon
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDormRequest {
  name?: string;
  description?: string;
  monthlyRent?: number;
  address?: string;
  latitude?: number;
  longitude?: number;
  amenities?: string[];
  totalRooms?: number;
  bedsPerRoom?: number;
  availableBeds?: number;
  coverPhotoUrl?: string;
}